Prometheus: Fix Prometheus endpoints in apache config
This updates the endpoints in the apache configuration for
Prometheus to correctly define the file used for http basic auth
to validate the admin user. The Prometheus endpoints restricted to
the admin user specified file for the authbasicprovider, but did
not provide the file used for validating the user. This adds the
file correctly

Change-Id: I8561281236fb1efa2e51af342e30314aae8e5285