Prevent resource deletion in Live Lab
The tempest utility that is configured here will take a snapshot.
After the snapashot and test run it will delete any resources that
are not in the snapshot.

If a tenant creates a resource outside of the test run it will also
be deleted.

This is something which is not desirable in a live lab.

This patch aims to provide control to the end-user to define how
to perform the final cleanup.

Change-Id: I1f7eb3ac5e9f9be08cb05a70bb1fb2d0f4b41382