[Open vSwitch] Remove auto_bridge_add support
We have two functionally identical places where we add bridges, one in
the neutron chart and one in the openvswitch chart.

It makes more sense to do it only in the neutron chart as that aligns
with the linux_bridge configuration and also is where the
bridge_mappings are specified.

Change-Id: I655380b021b89c3d93475febf7daca8f9d88cc54