Fluentbit: Update version, config util template
This updates fluentbit to version v0.14.2, which includes
the Modify plugin (required for trimming underscores from
systemd log fields, necessary for kibana visualization). This also
updates the fluentbit configuration util to allow for renaming
multiple entries in an event. This is required because the values
definition for a configuration section is defined as a map, and
does not supported multiple Rename directives

Change-Id: I05172e8236282a438587887f4a806cf35c4b6c68