Fluent-logging: Update kubernetes plugin test
This updates the kubernetes plugin test for fluent-logging to
search across all indices instead of the default logstash-* index
to account for custom indexes created for the events tagged with
the kubernetes plugin.

This also makes the search pattern for the tag more flexible to
account for any arbitrary number of prefixes and/or suffixes
added to the 'kube' tag as a result of any processing done in
fluentd.

Change-Id: Ib1a431cc8b2ca2cc143a8c8337b87f54f56d1029