Clean up the tempest.conf section of the values.yaml
While running chart internal to AT&T we found some
tempest.conf values that are not mapped to the correct section of the
tempest config.py.  This commmit is to address these changes:

1) tempest.auth.admin_domain_scope should be in
tempest.identity.admin_domain_scope
default is false and existing code is
setting the value to true so it is needed
https://github.com/openstack/tempest/blob/master/tempest/config.py#L109

2) tempest.auth.min_compute_nodes should be in
tempest.compute.min_compute_nodes but it is setting the default value so
this is not needed and I am dropping it
https://github.com/openstack/tempest/blob/14.0.0/tempest/config.py#L231

Also while learning about the chart we found that the admin_domain_scope
configmap-etc was mappinng the wrong data type so we removed the map
and set the admin_domain_scope  to true.

Change-Id: I52f115c3a1afee0fd58ef9422c6353ea1a93e814