Regroup OpenStack-Helm* gating under a folder
This move definitions of openstack-helm-infra into
a newly created zuul.d folder.

The advantage is to simplify readability of gating, and
makes it easier for contributors to step into the gating
of the openstack-helm-* projects.

- zuul.d/playbooks will contain all the playbooks used for gating
- zuul.d/nodesets.yaml contains all the specific nodesets
  required by OpenStack-Helm* projects
- zuul.d/project.yaml will be defined in each repo, and will
  contain the repo's pipelines information (so this repository's
  project.yaml only contains openstack-helm-infra pipelines)
- zuul.d/jobs.yaml will contain all the openstack-helm-*
  repositories jobs

This patch also introduces a first common 'lint' playbook
and 'openstack-helm-lint' job, showing how a job can be
re-used across repositories without requiring repetition of
job definition/plays in other repositories.

Change-Id: Id055ddac4da4971b1fb13ac075a7659369cd2b24