Use the openstack-helm-lint job from central location
This moves the job definitions into a single location, the zuul.d
folder, like it is done in the openstack-helm-infra repository.

The job playbooks and definitions of this repository will be moved
away to the openstack-helm-infra repository in a later commit.

This patch implements the first job reuse with the job
openstack-helm-lint from openstack-helm-infra's repository.

Change-Id: I233da952548d69f040771007455955131139e860
Depends-On: https://review.openstack.org/#/c/608662/