Prepare migration of base jobs to OSH-infra
Without this patch, openstack-helm-functional and
openstack-helm-multinode are a repetition of the
openstack-helm-infra-functional and openstack-helm-infra-multinode
jobs.

This is a problem, as multiple patches are required to change a base
job, and can lead to inconsistencies.

Instead the abstract jobs should be defined in only one place, the
openstack-helm-infra repositories. This will therefore be the location
of the openstack-helm project lint, functional, and multinode "base"
jobs.

From that point, openstack-helm-infra and openstack-helm can implement
their own job variants should this be required.

This patch fixes the link dependency between the two repositories by
ensuring all the OSH jobs are using their own local job definition
("-temp" jobs) during the migration.

Change-Id: I3d3193415ba65d58c5dd9bbb1fc515f5399edece
Depends-On: https://review.openstack.org/610482
Depends-On: https://review.openstack.org/610488
Depends-On: https://review.openstack.org/611239