Define OSH_PATH by default
OSH_PATH is not defined by default outside OpenStack's CI.

This is a problem if a user wants to run scripts manually on its
machine for local testing.

This fixes it by having, by default, the OSH_PATH defined
in the scripts using OSH relatively to current folder.

For user experience, the script returns to the same path after
running.

Change-Id: I915e7d3c945f2002a2008b2b033a2b7725320b17