Do not use OSH_INFRA_PATH in osh-infra
If OSH_INFRA_PATH is never used in the openstack-helm-infra repository,
as all the references are using relative paths.

The keystone script is not using a relative path, and relies on
OSH_INFRA_PATH to be defined to work.

This is a problem, because when it is not defined, the expected path
for ldap chart is /ldap, which is an incorrect path.

This fixes the problem by ensuring the path is relative.

Change-Id: I04a8d5c074b7c1e6fa66617bbb907f2ad4dcb3af