Nagios: Update host check mechanism
This updates the Nagios image to use a tag that includes a fix for
the service discovery mechanism used for updating host checks.
After moving the Nagios chart to either run in shared or host PID
namespaces, the service discovery mechanism no longer worked due
to the plugin attempting to restart PID 1 instead of determining
the appropriate PID to restart.

For reference, see:
https://review.gerrithub.io/#/c/att-comdev/nagios/+/432205/

Change-Id: Ie01c3a93dd109a9dc99cfac5d27991583546605a