[Ceph] Tunables for rgw buckets
Set rgw_override_bucket_index_max_shards to 8 (default: 0)

By default create 8 shards per a bucket with Ceph RagosGW. This allows
up to ~800k-1M objects to be in a bucket before seeing performance slow-
downs. The only downside to this change is that a directory listing for
a bucket may take slightly longer to finish.

Change-Id: I96c7ac81501a41d29927e102a6029bf432bd3d21