Add pre-fixes to the Selenium jobs and remove "|| true"
This adds xxx-job name prefixes to the Selenium jobs for consistency

This will also remove the "|| true" suffix that was added temporarily to
ensure the Kibana selenium job did not error. The fix for the issue
was merged so the quick fix is no longer needed and may prevent an
error when an issue actually occurs.
Change-Id: I16881974cbf618b31813964b17c090dbfe33fe51