Elastic Curator: Add basic action overrides for deployment jobs
This adds configuration overrides for a very basic Curator action
that should effectively be a no-op. This is to address periodic
failures seen in the osh-infra-aio-logging job that appear when
the run times coincide with Elastic Curator's cron schedule (every
six hours). This ensures curator actions are defined in cases
where this occurs

Change-Id: Ia2255ada2f32f21888bd4ca96df88496720fd0a5