Add preliminary support for standalone ironic
currently ironic chart is quite entangled with the presense of
other openstack services (Glance, Neutron, Swift).

Ironic is capable of running completely standalone, and while
the keystone and some neutron-related pieces are implemented as jobs
and can be turned off in manifests and dependencies sections of values,
others are scripts running as initContainers and are not the easy to
switch off.

This patch adds more key/values to the bootstrap key,
which allows to turn off Neutron-, Swift- and Glance-related pieces
while keeping possibility that some other networking, image or
object_store actions appropriate for standalone case may be needed.

Change-Id: Icccbdbce81ca350042f33f5e86bb942064839267