Improves ceph-provisioner storageclasses
Currently configuring storage classes is static. If
we want to add another storageclass for an exmaple
for harddisk then we have to create another yaml
along with update in .Values.yaml. Using this change
we can create as many storageclasses we want. Change
will basically iterate in loop and apply the values
present in .Values

At the same time we have to update the changes in
.Values.storageclass.rbd paramert values in all
applicable places

Change-Id: I3207f047c3f8b1a33dfb97c87b9ea7120b54c4b4
Story: ceph-provisioners-improvement
Signed-off-by: Kranthi Guttikonda <kranthi.guttikonda@b-yond.com>
Signed-off-by: kranthikirang <kranthi.guttikonda@b-yond.com>