Cleanup specs folder
Specs are not ordered currently, and every rst file inside the
specs folder is included in the TOC tree, but manually.

This is a problem as:
- the current readability of the specs was reduced due to inclusion
  of non-specs files
- the process of writing a spec was more tedious, due to the
  update of the specs/index.rst.

This fixes it by removing the extra files included by mistake in
the middle of the specs (the template for spec writing, and
the specs purpose/process), and automatically load all the
remaining files using a glob.

The content of the files removed is not lost: The template was
simply renamed COPYME to clearly state a spec writer should
copy the file (and will understand it needs to be named .rst)
with the other files present. The specs process/purpose is
now part of the main page of specs, which therefore doesn't need
extra including.

Change-Id: I8aa15c8a8f2d8b3ffb764c3fb2411eb27477d0b6