RabbitMQ: Make clustering robust
This PS improves the robustnes of the RabbitMQ clustering logic
to support reforming the cluster following recreation of all pods,
and wait for the cluster to fully form before continuing in case
of an upgrade.

This ability was lost with the introduction of the following PS,
which prevented reformation of the cluster from scratch.
 * https://review.openstack.org/#/c/637337/

Change-Id: I99d32fbd3c56dde492717a7850b61001fa8f7fb5
Signed-off-by: Pete Birley <pete@port.direct>