Elasticsearch: Remove insertion of test data in helm tests
This removes the insertion of test dummy data and the following
query for it from the Elasticsearch helm tests.  Upon upgrades,
it's possible for Elasticsearch to refuse the direct insertion of
data due to shard reallocation and due to full bulk endpoint
queues. These refusals should not be seen as test failures

Change-Id: Id53d53a7aa2b58e64932d50ca3e7a4fb1141bb3a