MariaDB Leader Election Enhancement
This patch set enhances the current MariaDB leader election algorithm
by adding a tiebreaker in case multiple nodes have the same sequence
number which happens to be the highest sequence number in the MariaDB
cluster. The tiebreaker is to pick the node which has the lowest node
number in the hostname.

This patch also changes the 409 status code reporting inside the
safe_update_configmap() function to be a warning instead of an error,
as it is indicative of a collision scenario rather than an actual
error.

Change-Id: Ifdff0250b520bb1972d79de94a491e566ed43997