Revert "readOnlyRootFilesystem: true for openvswitch chart"
I believe when we have set the readOnly flag at pod without HTK functionality the changes were not reflected. That is why it passed the gate.

Later with HTK functionality the gates never passed and I have tested that in various ways and finally I had to unset the readOnly flag

This reverts commit 598040bea05737ea1ee2460ba8675ed7c061e63a.

Change-Id: Icf8d3cc60045926ab60b9735ee1e8202c15df9d5