Fix prometheus metrics gathering in postrun job
This adds '|| true' to the curl command for gathering metrics from
prometheus exporters in the postrun job. After the move to
minikube for single node jobs, the headless services for the
kubernetes components no longer work as intended. The addition of
'|| true' allows the post run job to continue through the list
of services tied to the prometheus exporters without the task
failing outright

Change-Id: I56f0f56b799c3df9b2bd66a2c2044d71473606e3